fix: follow symlinks when grepping for desktop entries
When running browsers installed via the Nix home-manager, the browser
.desktop files will be found within `~/.nix-profile/share/applications/`
and `~/.nix-profile/share` will be in `$XDG_DATA_DIRS`. The
`~/.nix-profile/share` directory is a symlink, and `grep -r` will follow
that successfully, but then the `applications/firefox.desktop` within
will also be a symlink. A `grep -r` won't follow that, so we have to use
`grep -R` to follow all symlinks to actually find the desktop entries.

For example, without this fix, I see the following when running
`linkquisition https://google.com`.

    failed to call grep for determining a .desktop entry for firefox -P Personal --class=firefox-personal %U: exit status 1
    failed to call grep for determining a .desktop entry for google-chrome-stable --user-data-dir="/home/cluther/.config/google-chrome-personal" %U: exit status 1

With this change, linkquisition is able to auto-detect all of the
browser entries I have, and launch into them without error.